Improve field boss monitoring and robustness
Several fixes and improvements around field boss handling and related services:

- Unify channel ID type: change AliveChannels keys from short to int across World and Game components.
- ChannelService: add monitoringBossFields to prevent duplicate lifetime monitors; MonitorExtendedBossLifetimeAsync now takes targetMapId, guards against LastDamageTick==0, uses logger, and removes the map from monitoring on exit.
- WorldServer: make MonitorFieldBossLifetimeAsync cancellable by accepting a CancellationToken and using cancellable Task.Delay; suppress OperationCanceledException during shutdown.
- GameSession: handle RpcException when fetching active field bosses and fix grouping/metadata lookup to correctly build boss groups and send world map data.
- FieldBossUtil: compute next spawn by calculating cycles instead of looping to efficiently advance to the next occurrence.
- NoticeCommand: adjust SetHandler generic parameters and early-return when an invalid StringCode is provided.

These changes improve correctness, prevent duplicate monitoring tasks, and add resiliency during RPC failures and server shutdown.

+++ Maple2.Server.World/Containers/FieldBossManager.cs
@@ -12,7 +12,7 @@ public class FieldBossManager : IDisposable {
     public required ChannelClientLookup ChannelClients { get; init; }
 
     public readonly FieldBoss Boss;
-    public readonly ConcurrentDictionary<short, byte> AliveChannels = new();
+    public readonly ConcurrentDictionary<int, byte> AliveChannels = new();
 
     public FieldBossManager(FieldBossMetadata metadata, int id, long endTick, long nextSpawnTimestamp) {
         Boss = new FieldBoss(metadata, id) {
@@ -21,7 +21,7 @@ public class FieldBossManager : IDisposable {
         };
     }
 
-    public void RemoveChannel(short channel) => AliveChannels.TryRemove(channel, out _);
+    public void RemoveChannel(int channel) => AliveChannels.TryRemove(channel, out _);
 
     public void Announce() {
         foreach ((int channelId, ChannelClient channelClient) in ChannelClients) {
@@ -35,7 +35,7 @@ public class FieldBossManager : IDisposable {
                     },
                 });
 
-                AliveChannels.TryAdd((short) channelId, 0);
+                AliveChannels.TryAdd(channelId, 0);
             } catch (RpcException rpcException) {
                 if (rpcException.StatusCode == StatusCode.Unavailable) {
                     Log.Warning("Channel {Channel} unavailable when announcing field boss {BossId}", channelId, Boss.MetadataId);
